Fix non-static method call to UserProgress::hasCompletedExam()
- Updated `HomeController::index()` to instantiate `UserProgress` and set `user_id` before calling the `hasCompletedExam` method.
- Removed static call to a non-static method to resolve the "Non-static method cannot be called statically" error.

The `hasCompletedExam` method in the `UserProgress` model is an instance method that relies on the `user_id` property. Previously, it was incorrectly called statically, causing an exception. By creating an instance of `UserProgress` and setting the `user_id`, the method can be invoked correctly, allowing the Home view to determine the progress status of exams accurately.